[TAN-4816] Moved mailers back and defined preview on the mailer class
