+++ back/engines/free/email_campaigns/app/models/email_campaigns/campaigns/comment_on_idea_you_follow.rb
@@ -53,10 +53,6 @@ module EmailCampaigns
       CommentOnIdeaYouFollowMailer
     end
 
-    def preview_class
-      CommentOnIdeaYouFollowMailerPreview
-    end
-
     def self.recipient_role_multiloc_key
       'email_campaigns.admin_labels.recipient_role.registered_users'
     end
