+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/comment_on_idea_you_follow_mailer_preview.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module EmailCampaigns
+  class CommentOnIdeaYouFollowMailerPreview < ActionMailer::Preview
+    include EmailCampaigns::MailerPreviewRecipient
+
+    def campaign_mail
+      EmailCampaigns::CommentOnIdeaYouFollowMailer.preview_email(
+        campaign: Campaigns::CommentOnIdeaYouFollow.first,
+        recipient: recipient_user
+      )
+    end
+  end
+end
